UINV-371 Move Export btn from settings
