+++ src/invoices/VoucherExport/ExportVoucherButton/index.js
@@ -1,0 +1,1 @@
+export { default as ExportVoucherButton } from './ExportVoucherButton';
