+++ src/invoices/VoucherExport/index.js
@@ -1,1 +1,2 @@
 export { default as VoucherExport } from './VoucherExport';
+export { ExportVoucherButton } from './ExportVoucherButton';
