--- src/settings/BatchGroupConfigurationSettings/ExportVoucherButton/index.js
@@ -1,1 +1,0 @@
-export { default } from './ExportVoucherButton';
